Basic html for the daily calander
TEMPLATE.
This is just a template version of the html for the daily calendar. I plan on reducing all of the redundancy but need help with the cards template for data.
TODO:
I will update the the html to be more of a optomized version. I will then update the style to match more of a "stack of cards"