Full improvement of daily calendar
Using the functionality of the template file I had created yesterday, I created the necessary JS to improve upon it.

TODO:
I still need to create the CSS file for this feature however i already have a decent idea on how to do it so it should be fast. I will then improve styling of my code to fit the repo better and pass all test cases.