CSS/functionality update
I removed the button on the right and left because it felt redundant as I already include that feature. added the css template(will update later to fit the figma design)
TODO:
Refactor code to fit repo design style.
add feature to click on the card to open it up as well and tasks complete on the day card